requirements.txt
django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2
psycopg2-binary==2.9
python-decouple==3.8
django-cors-headers==4.0
social-auth-app-django==5.0